pyproject.toml
[project]
name = "weather-ingester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "notebook>=7.5.4",
    "pgcli>=4.4.0",
]
